feat: implement differential parity testing & fix bignum edge cases
- Implement JvmBigNumComparisonTest against JDK math baseline
- Implement KmpBigNumComparisonTest against Ionspin Multiplatform bignum
- Fix potential 64-bit Knuth division overflow inside BigInteger
- Fix BigDecimal.toString() layout formatting for zero with positive scales
- Add new Mathematics category and examples to root README
- Add detailed Parity Test Status Table to README.md
- Correct broken/outdated documentation links in README.md
- Track kotlinx.benchmark integration as a project gap in GAPS.md
- Update library embedded AI-skills metadata for BigDecimal & BigInteger
- Update CHANGELOG.md under Unreleased with additions and fixes
- Clean up misplaced transient planning files from docs/research/

+++ types/src/commonMain/kotlin/io/github/aughtone/types/number/BigInteger.kt
@@ -572,7 +572,7 @@ class BigInteger internal constructor(
                 var qHat = ((uJn shl 32) or uJn1) / vN1
                 var rHat = ((uJn shl 32) or uJn1) % vN1
                 
-                while (qHat >= 4294967296L || qHat * vN2 > (rHat shl 32) + uJn2) {
+                while (qHat >= 4294967296L || (qHat.toULong() * vN2.toULong()) > ((rHat.toULong() shl 32) + uJn2.toULong())) {
                     qHat--
                     rHat += vN1
                     if (rHat >= 4294967296L) break
